Split image fading code from fading transition code so scene controller doesn't need a reference to a transition.
Also removed fade to image settings scriptable object in favour of an approach where settings are set on the image component directly.
