+++ Assets/Scenes/MainMenu.unity
@@ -1566,9 +1566,9 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: a791a1928c03049c78bb7a707a30453d, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  _image: {fileID: 1961069501}
+  _imageFader: {fileID: 1961069503}
   _shouldFadeInOnAwake: 0
-  _fadeToImageSettings: {fileID: 11400000, guid: 24785bd9ba2a14df89de403b8a67a10f, type: 2}
+  _fadeTime: 1
   _onTransitionOutComplete:
     m_PersistentCalls:
       m_Calls:
@@ -6976,9 +6976,9 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: a791a1928c03049c78bb7a707a30453d, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  _image: {fileID: 1961069501}
+  _imageFader: {fileID: 1961069503}
   _shouldFadeInOnAwake: 0
-  _fadeToImageSettings: {fileID: 11400000, guid: 24785bd9ba2a14df89de403b8a67a10f, type: 2}
+  _fadeTime: 1
   _onTransitionOutComplete:
     m_PersistentCalls:
       m_Calls:
@@ -8692,13 +8692,14 @@ GameObject:
   - component: {fileID: 1961069500}
   - component: {fileID: 1961069502}
   - component: {fileID: 1961069501}
+  - component: {fileID: 1961069503}
   m_Layer: 5
   m_Name: BlackScreen
   m_TagString: Untagged
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
   m_StaticEditorFlags: 0
-  m_IsActive: 1
+  m_IsActive: 0
 --- !u!224 &1961069500
 RectTransform:
   m_ObjectHideFlags: 0
@@ -8725,7 +8726,7 @@ MonoBehaviour:
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
   m_GameObject: {fileID: 1961069499}
-  m_Enabled: 0
+  m_Enabled: 1
   m_EditorHideFlags: 0
   m_Script: {fileID: 11500000, guid: fe87c0e1cc204ed48ad3b37840f39efc, type: 3}
   m_Name: 
@@ -8756,6 +8757,18 @@ CanvasRenderer:
   m_PrefabAsset: {fileID: 0}
   m_GameObject: {fileID: 1961069499}
   m_CullTransparentMesh: 1
+--- !u!114 &1961069503
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 1961069499}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: c516981a0c274483bb82bc6d981a84a1, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
 --- !u!1 &1970614773
 GameObject:
   m_ObjectHideFlags: 0
