--- Assets/ScriptableObjects/GameSettings.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: ff10ec1f0484049e7b21b468fa0842d0
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
