--- Assets/ScriptableObjects/GameSettings/FadeToBlack.asset.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 24785bd9ba2a14df89de403b8a67a10f
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 11400000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
