+++ Assets/Scripts/BGScenes/ImageFader.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 793e414816ca440bebe271163a60ecd4
+guid: c516981a0c274483bb82bc6d981a84a1
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
